+++ openstack/blockstorage/extensions/schedulerstats/list.go
@@ -4,10 +4,51 @@ import (
 	"encoding/json"
 	"math"
 
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
 	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
-// Capabilities represents the information of an individual StoragePool.
+type ListOpts struct {
+	// ID of the tenant to look up storage pools for.
+	TenantID string `q:"tenant_id"`
+	// Whether to list extended details.
+	Detail bool `q:"detail"`
+}
+
+func List(client *golangsdk.ServiceClient, opts ListOpts) pagination.Pager {
+	query, err := golangsdk.BuildQueryString(opts)
+	if err != nil {
+		return pagination.Pager{Err: err}
+	}
+
+	return pagination.NewPager(client, client.ServiceURL("scheduler-stats", "get_pools")+query.String(), func(r pagination.PageResult) pagination.Page {
+		return StoragePoolPage{pagination.SinglePageBase(r)}
+	})
+}
+
+type StoragePoolPage struct {
+	pagination.SinglePageBase
+}
+
+func (page StoragePoolPage) IsEmpty() (bool, error) {
+	va, err := ExtractStoragePools(page)
+	return len(va) == 0, err
+}
+
+func ExtractStoragePools(p pagination.Page) ([]StoragePool, error) {
+	var res struct {
+		StoragePools []StoragePool `json:"pools"`
+	}
+	err := extract.Into(p.(StoragePoolPage).Body, &res)
+	return res.StoragePools, err
+}
+
+type StoragePool struct {
+	Name         string       `json:"name"`
+	Capabilities Capabilities `json:"capabilities"`
+}
+
 type Capabilities struct {
 	// The following fields should be present in all storage drivers.
 	DriverVersion     string  `json:"driver_version"`
@@ -16,7 +57,6 @@ type Capabilities struct {
 	TotalCapacityGB   float64 `json:"-"`
 	VendorName        string  `json:"vendor_name"`
 	VolumeBackendName string  `json:"volume_backend_name"`
-
 	// The following fields are optional and may have empty values depending
 	// on the storage driver in use.
 	ReservedPercentage       int64   `json:"reserved_percentage"`
@@ -31,13 +71,6 @@ type Capabilities struct {
 	GoodnessFuction          string  `json:"goodness_function"`
 	Mutliattach              bool    `json:"multiattach"`
 	SparseCopyVolume         bool    `json:"sparse_copy_volume"`
-}
-
-// StoragePool represents an individual StoragePool retrieved from the
-// schedulerstats API.
-type StoragePool struct {
-	Name         string       `json:"name"`
-	Capabilities Capabilities `json:"capabilities"`
 }
 
 func (r *Capabilities) UnmarshalJSON(b []byte) error {
@@ -73,26 +106,4 @@ func (r *Capabilities) UnmarshalJSON(b []byte) error {
 	r.TotalCapacityGB = parseCapacity(s.TotalCapacityGB)
 
 	return nil
-}
-
-// StoragePoolPage is a single page of all List results.
-type StoragePoolPage struct {
-	pagination.SinglePageBase
-}
-
-// IsEmpty satisfies the IsEmpty method of the Page interface. It returns true
-// if a List contains no results.
-func (page StoragePoolPage) IsEmpty() (bool, error) {
-	va, err := ExtractStoragePools(page)
-	return len(va) == 0, err
-}
-
-// ExtractStoragePools takes a List result and extracts the collection of
-// StoragePools returned by the API.
-func ExtractStoragePools(p pagination.Page) ([]StoragePool, error) {
-	var res struct {
-		StoragePools []StoragePool `json:"pools"`
-	}
-	err := (p.(StoragePoolPage)).ExtractInto(&res)
-	return res.StoragePools, err
 }
